Move convert_to_column_reference and TABLE_REFERECE
